Resolve S13P11A601-47 "Feature/"

See merge request s13-webmobile1-sub1/S13P11A601!141

+++ AI/src/specification/chat_router.py
@@ -33,7 +33,7 @@ async def chat_stream(
     intent: ChatIntent,
     project_id: int,
     message: str,
-    time: int,
+    time: float,
     feature_id: int = None,
     title: str = None,
     field: str = None,
